// assets/script/proto/game_common_player.ts
export interface PlayerExitRoomRes {
  /** 退出房间响应状态 */
  status: RoomActionAcceptStatus;
  /** 当前最新预约动作 */
  activePendingAction: PendingRoomAction;
}
// ...
export interface PlayerSwitchRoomRes {
  status: RoomActionAcceptStatus;
  activePendingAction: PendingRoomAction;
}
// ...
export interface PlayerSitDownRes {
  seat: number;
  /** 等待玩家准备倒计时的结束时间，单位秒。为0表示没有开启倒计时 */
  waitReadyExpiredTime: number;
}
// ...
export interface PlayerStandUpRes {
}
// ...
export interface PlayerReadyGameRes {
  ready: boolean;
  /** 等待玩家准备倒计时的结束时间，单位秒。为0表示没有开启倒计时 */
  waitReadyExpiredTime: number;
}
// ...
export const PlayerExitRoomRes = {
  encode(message: PlayerExitRoomRes, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.status !== 0) {
      writer.uint32(8).int32(message.status);
    }
    if (message.activePendingAction !== 0) {
      writer.uint32(16).int32(message.activePendingAction);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): PlayerExitRoomRes {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePlayerExitRoomRes();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.status = reader.int32() as any;
          continue;
        case 2:
          if (tag !== 16) {
            break;
          }

          message.activePendingAction = reader.int32() as any;
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): PlayerExitRoomRes {
    return {
      status: isSet(object.status) ? roomActionAcceptStatusFromJSON(object.status) : 0,
      activePendingAction: isSet(object.activePendingAction)
        ? pendingRoomActionFromJSON(object.activePendingAction)
        : 0,
    };
  },

  toJSON(message: PlayerExitRoomRes): unknown {
    const obj: any = {};
    if (message.status !== 0) {
      obj.status = roomActionAcceptStatusToJSON(message.status);
    }
    if (message.activePendingAction !== 0) {
      obj.activePendingAction = pendingRoomActionToJSON(message.activePendingAction);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<PlayerExitRoomRes>, I>>(base?: I): PlayerExitRoomRes {
    return PlayerExitRoomRes.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<PlayerExitRoomRes>, I>>(object: I): PlayerExitRoomRes {
    const message = createBasePlayerExitRoomRes();
    message.status = object.status ?? 0;
    message.activePendingAction = object.activePendingAction ?? 0;
    return message;
  },
};
// ...
export const PlayerSwitchRoomRes = {
  encode(message: PlayerSwitchRoomRes, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.status !== 0) {
      writer.uint32(8).int32(message.status);
    }
    if (message.activePendingAction !== 0) {
      writer.uint32(16).int32(message.activePendingAction);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): PlayerSwitchRoomRes {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePlayerSwitchRoomRes();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.status = reader.int32() as any;
          continue;
        case 2:
          if (tag !== 16) {
            break;
          }

          message.activePendingAction = reader.int32() as any;
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): PlayerSwitchRoomRes {
    return {
      status: isSet(object.status) ? roomActionAcceptStatusFromJSON(object.status) : 0,
      activePendingAction: isSet(object.activePendingAction)
        ? pendingRoomActionFromJSON(object.activePendingAction)
        : 0,
    };
  },

  toJSON(message: PlayerSwitchRoomRes): unknown {
    const obj: any = {};
    if (message.status !== 0) {
      obj.status = roomActionAcceptStatusToJSON(message.status);
    }
    if (message.activePendingAction !== 0) {
      obj.activePendingAction = pendingRoomActionToJSON(message.activePendingAction);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<PlayerSwitchRoomRes>, I>>(base?: I): PlayerSwitchRoomRes {
    return PlayerSwitchRoomRes.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<PlayerSwitchRoomRes>, I>>(object: I): PlayerSwitchRoomRes {
    const message = createBasePlayerSwitchRoomRes();
    message.status = object.status ?? 0;
    message.activePendingAction = object.activePendingAction ?? 0;
    return message;
  },
};
// ...
export const PlayerSitDownRes = {
  encode(message: PlayerSitDownRes, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.seat !== 0) {
      writer.uint32(8).int32(message.seat);
    }
    if (message.waitReadyExpiredTime !== 0) {
      writer.uint32(32).int64(message.waitReadyExpiredTime);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): PlayerSitDownRes {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePlayerSitDownRes();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.seat = reader.int32();
          continue;
        case 4:
          if (tag !== 32) {
            break;
          }

          message.waitReadyExpiredTime = longToNumber(reader.int64() as Long);
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): PlayerSitDownRes {
    return {
      seat: isSet(object.seat) ? globalThis.Number(object.seat) : 0,
      waitReadyExpiredTime: isSet(object.waitReadyExpiredTime) ? globalThis.Number(object.waitReadyExpiredTime) : 0,
    };
  },

  toJSON(message: PlayerSitDownRes): unknown {
    const obj: any = {};
    if (message.seat !== 0) {
      obj.seat = Math.round(message.seat);
    }
    if (message.waitReadyExpiredTime !== 0) {
      obj.waitReadyExpiredTime = Math.round(message.waitReadyExpiredTime);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<PlayerSitDownRes>, I>>(base?: I): PlayerSitDownRes {
    return PlayerSitDownRes.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<PlayerSitDownRes>, I>>(object: I): PlayerSitDownRes {
    const message = createBasePlayerSitDownRes();
    message.seat = object.seat ?? 0;
    message.waitReadyExpiredTime = object.waitReadyExpiredTime ?? 0;
    return message;
  },
};
// ...
export const PlayerStandUpRes = {
  encode(_: PlayerStandUpRes, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): PlayerStandUpRes {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePlayerStandUpRes();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(_: any): PlayerStandUpRes {
    return {};
  },

  toJSON(_: PlayerStandUpRes): unknown {
    const obj: any = {};
    return obj;
  },

  create<I extends Exact<DeepPartial<PlayerStandUpRes>, I>>(base?: I): PlayerStandUpRes {
    return PlayerStandUpRes.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<PlayerStandUpRes>, I>>(_: I): PlayerStandUpRes {
    const message = createBasePlayerStandUpRes();
    return message;
  },
};
// ...
export const PlayerReadyGameRes = {
  encode(message: PlayerReadyGameRes, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.ready !== false) {
      writer.uint32(8).bool(message.ready);
    }
    if (message.waitReadyExpiredTime !== 0) {
      writer.uint32(16).int64(message.waitReadyExpiredTime);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): PlayerReadyGameRes {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePlayerReadyGameRes();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.ready = reader.bool();
          continue;
        case 2:
          if (tag !== 16) {
            break;
          }

          message.waitReadyExpiredTime = longToNumber(reader.int64() as Long);
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): PlayerReadyGameRes {
    return {
      ready: isSet(object.ready) ? globalThis.Boolean(object.ready) : false,
      waitReadyExpiredTime: isSet(object.waitReadyExpiredTime) ? globalThis.Number(object.waitReadyExpiredTime) : 0,
    };
  },

  toJSON(message: PlayerReadyGameRes): unknown {
    const obj: any = {};
    if (message.ready !== false) {
      obj.ready = message.ready;
    }
    if (message.waitReadyExpiredTime !== 0) {
      obj.waitReadyExpiredTime = Math.round(message.waitReadyExpiredTime);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<PlayerReadyGameRes>, I>>(base?: I): PlayerReadyGameRes {
    return PlayerReadyGameRes.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<PlayerReadyGameRes>, I>>(object: I): PlayerReadyGameRes {
    const message = createBasePlayerReadyGameRes();
    message.ready = object.ready ?? false;
    message.waitReadyExpiredTime = object.waitReadyExpiredTime ?? 0;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };